Rethink exception-handling in _prep_response.
Quit catching TypeError; requests catches ValueError (which is the only thing likely to be raised from loads()--you'd have to have the response magically turn into some non-string Python type to get a TypeError) and returns None instead.

We can return a more specific exception later, backward-compatibly.

Be explicit in the exception's error message that it was ES that spit out the invalid JSON.