Decorate a few more methods.
The nice thing about this approach is that it puts all the ugly stuff (and it's not even that ugly) in the lib. Callers don't have to bother about it; for them, it's just as if we were passing **kwargs straight through (plus the es_ thing, which is nice for them because they'll never have to rewrite their bleeding-edge code).