Always use the singular spelling for "index" and "doc_type" kwargs, and make up the difference with documentation.
It's not always obvious whether an ES API allows for multiple indexes. This was leading me to have to look aside to the docs to determine whether the kwarg was called "index" or "indexes". Let's just use the singular everywhere, and then the common case--a single index--won't result in any doc lookups.

Same with doctypes.