Compatibility with Python stdlib / simplejson<2.1
pyelasticsearch.py has two exception handlers which catch json.JSONDecodeError.
Unfortunately, that was only added in simplejson 2.1.0 and because the import
uses the requests.compat.json failover mechanism it's easy to install and use
pyelasticsearch without having installed simplejson 2.1.0, which will produce
unexpected AttributeErrors.

A quick test shows that the stdlib json module in Python 2.6, 2.7 and 3.2 lacks
the JSONDecodeError.