Remove percolator() method, and tweak docstrings.
Sitting down to merge this in, I had second thoughts about percolator(). I think georgi0u was right and that people should just use index() instead. Otherwise, we have to add not only this method (which I'd call "put_percolator()") but also a symmetric one to delete percolators. And that's just a lot of extra routines that don't any knowledge except the '_percolator' string. And then what it we want to count percolators or something? So I think I'll leave this out for now. We can always add it later.