Fix problem with fluctuating scores in tests
The tests were such that they only passed on a specific version of
Elasticsearch because they compared against specific scores in the
results.

Since we don't actually care about what the scores are, we do this
sleight-of-hand with WHATEVER which happily compares positively with
everything.

This fixes the tests so everything passes in Elasticsearch 0.90.10.

Fixes #107