Fix tests to pass with ES 0.90 through ES 1.3.4
This fixes the tests to pass with Elasticsearch 0.90 through
Elasticsearch 1.3.4.

Some of the fixes use eq_one_of which allows us to run multiple test
conditions and as long as one of them passes, we're fine. This allows us
to handle the change in HTTP responses from Elasticsearch between 0.90
and 1.0+.