Add support for paths rooted beyond /
It is possible to put the ElasticSearch api behind a proxy with a
different root path than /. Old code assumed / was used everywhere.
Update things to accept a longer path to support ElasticSearch servers
proxied in this manner.

Note that because we rely on elasticsearch lib's Transport class we
can't currently allow different paths for each host. You are still
limited to having all hosts proxied behind the same path.